feat (utils): add module with helper functions
