+++ .gitignore
@@ -7,3 +7,4 @@ env/*
 .env
 docs/design-documentation.pdf
 docs/design-documentation.html
+utils/__pycache__
